requirements.txt
pyocclient==0.4
tqdm==4.19.5
Pillow==5.0.0
